Slight refactoring
The function replace_data_dispatch() has been renamed and is now called
unschedule_data_dispatch().
The difference is that now it is only responsible for finding and
removing the data dispatch event from the event queue, and returning it
to the data queue. The insertion of the voice dispatch event is made
inside enter_the_server().